Fix ale_web_client image tag to cloud4things organization
When build the container for ale_web_client, the script is specifying to
use the image mvpgomes/ale_web_client. However this image does not exist
in the remote repo and the creation will fail if there is no image in
the local env. This commit fixes the run script to pull the image from
the correct repository.